Add Image poro and test

+++ spec/requests/api/v1/city/forecast_spec.rb
@@ -22,7 +22,7 @@ RSpec.describe 'the city weather forecast endpoint' do
         expect(current[:sunset].to_datetime).to be_a(DateTime)
         expect(current[:feels_like]).to be_a(Float)
         expect(current[:humidity]).to be_a(Float)
-        expect(current[:uvi]).to be_a(Integer)
+        expect(current[:uvi]).to be_a(Numeric)
         expect(current[:visibility]).to be_a(Integer)
         expect(current[:conditions]).to be_a(String)
         expect(current[:icon]).to be_a(String)
